requirements.txt
alembic==1.0.6
blinker==1.4
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==1.0.2
Flask-HTTPAuth==3.2.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
httpie==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
pkg-resources==0.0.0
Pygments==2.3.1
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
requests==2.21.0
six==1.12.0
SQLAlchemy==1.2.15
urllib3==1.24.1
Werkzeug==0.14.1
WTForms==2.2.1
